+++ src/character.cpp
@@ -1286,13 +1286,13 @@ units::mass Character::weight_capacity() const
     ret += get_str() * 4_kilogram;
     ret *= mutation_value( "weight_capacity_modifier" );
 
-    int worn_weight_bonus = 0;
+    float worn_weight_bonus = 0;
     for( const item &it : worn ) {
         ret *= it.get_weight_capacity_modifier();
         worn_weight_bonus += it.get_weight_capacity_bonus();
     }
 
-    int bio_weight_bonus = 0;
+    float bio_weight_bonus = 0;
     for( const bionic &bio : *my_bionics ) {
         ret *= bio.info().weight_capacity_modifier;
         bio_weight_bonus += bio.info().weight_capacity_bonus;
