+++ src/item.h
@@ -1478,7 +1478,7 @@ class item : public visitable<item>
          * Returns the weight capacity bonus (@ref islot_armor::weight_capacity_modifier) that this item provides when worn.
          * For non-armor it returns 1. The bonus is added to the total weight capacity of the character that wears the item.
          */
-        int get_weight_capacity_bonus() const;
+        float get_weight_capacity_bonus() const;
         /**
          * Returns the resistance to environmental effects (@ref islot_armor::env_resist) that this
          * item provides when worn. See @ref player::get_env_resist. Higher values are better.
